Hyperlink colour/structure.
As a colour blind user I found the difference between dark and mid to subtle when I loaded up the change. Yet I found the dark to light to brash. So I stuck with dark to mid but with the added underline hint on mouse-over.

Introduced some style inheritence to reduce the duplication.