TimePicker now falls back to previous valid value
Changes made to force behavior of TimePicker to more closely follow the
behavior of the DatePicker control. When there is previously a good
value selected for the time and the user then enters an invalid time,
the time will fall back to the previous good value. The time parsing has
also been opened up to allow more formats.