Fixes #945 (#1141)
The issue is the PopupOnLoaded method that was setting the IsOpen property directly on the Popup. This would lock the popup in an opened state when it was hidden by the tab control. When coming back to the popup box the popup would not properly trigger the IpOpen change from the trigger preventing it from showing up again, effectively killing all of the contents in the popup