New Scrollbar Style (#1270)
* Added MaterialDesignOutlinedButton Style

Basically get the Flat Design and added a `BorderThickness` = 1, and changed the `RippleEffect` color. Also changed the background when it is pressed to not see the gray color, as in
https://material.io/design/components/buttons.html#outlined-button

* Added corner radius support

* Added New Scrollbar Style

* Remove custom brushes and Make it default

Removed the custom brushed in order to use the some of the already defined.
Also changed the default style for the scrollbar to this one so there is no need to change it everywhere.

* Cleaned up scroll bar style

Moved back to only having a single style and toggling values with attached properties.
Added attached properties for thumb corner radius, height, and width.
Renamed minimal style to match other styles in the library.