[Snackbar 1/3] Improve duplicate handling
Renamed IgnoreDuplicate to ShowAlways.
Added MessageExpired method.
Added appropriate Equals method.
Added autogenerated GetHashCode method.

ShowAlways more closely represented what that property actually does. IgnoreDuplicate is not very telling in the context of a SnachkBarMessageQueueItem.
MessageExpired serves as a helper to determine if a message already exceeded it's duration.
Equals and HashCode simplify comparisons of SnackBarMessageQueueItems, especially useful in the determining if a message is a duplicate.