Refactor fix for 2689 to use attached properties (#2762)
* Refactor fix for 2689 to use attached properties

Added attached properties for UniformCornerRadius and Style. These properties are inherited and can thus be used to manipulate a Card located inside of another UIElement (eg. the Flipper).

* Make CardAssist.CardStyle lenient towards unaccepted style

Fallback to default(Style) if the TargetType is not Card.

* Renamed CardAssist and improved guard in CardStyle setter (of attached property)

You mentioned on stream that fact that derived types would not work for the CardStyle because of the guard which is completely correct. So I loosened up the requirements by using IsAssignableFrom() instead.

* Fixed broken Flipper (UI) tests

Rename operation did not pick up the string literal in these tests