Allow "half" rating selection (#2764)
* RatingBar.Value changed to double

Value is now set as a double that is coerced to a multiple of 0.5 within the Min/Max bounds.

The colors are currently hardcoded, this needs to be addressed. Unit tests should also be added for the coercion.

* Added unit test for coercion

* Changed [Theory] to [StaTheory] to make unit tests run

* Fixed lower bound issue

Fixed issue where the first rating button could not be halfway selected.

It is worth noting that valid values are actually values between Min-1 and Max (in multiples of 0.5)

* POC for MultiValueConverter approach

* Opt-in for fractional values

Refactored into opt-in for fractional values. Fractional increments are now also configurable via the ValueIncrements DP.

* opt-in for PreviewValue added (no visualization)

* Moved multiconverter into nested class

* Added preview "tooltip"

Positioning of tooltip still missing

* Correct placement of value preview "tooltip"

* Added unit tests for transform converters

* Marking RatingBarButton.IsWithinSelectedValue as obsolete

Co-authored-by: Kevin Bost <kitokeboo@gmail.com>