Added BorderThickness attached DP for outlined date/time pickers (#2805)
* Added BorderThickness attached DP for outlined date/time pickers

* Added UI tests

* Fixed typos in code comment

* Increased delay in tests

Tests are running green locally, so I presume it has something to do with the delay being too short

* Fixed UI tests

Initial mouse position was on the pickers, so I added a dummy button to initially move the mouse position away from the pickers.

* Updated UI tests to test validation error scenario

Ideally I would like XAMLTest to allow me to set a ValidationError "on the fly" if possible. For now, tests are updated with a real ValidationRule and invalid values are set to provoke the validation error.