New feature - PasswordBox styles with "reveal password" functionality (#2828)
* Initial PasswordBox "revealed" styles

There are still issues with the positioning of the floating hint when password is revealed but empty; the hint should float up, but it does not.

* Fixed alignment and initial wiring up

Floating hint still misplaced when password is empty and revealed

* Showcasing style in sample app + alignment of reveal button

Reveal button aligns with the "clear text" button for consistency. Although I think the "clear text" button is floating a little too high in 2 of the styles (but that is the same for the normal PasswordBox and TextBox styles)

* Showcase that password can still be set from XAML

* Change PasswordChanged event handler to weak event pattern

Avoid leaking memory by replacing the regular event handler with a weak event handler.

* Updated showcasing in the demo tool with a bit more stuff

* Fixed CA warning

* Added UI test for 3-way binding

* [Icon update detected by Github Action]. Auto generated pull request. (#2830)

Co-authored-by: Material Design Service Account <mdixsa@outlook.com>

* Cursor fix (#2832)

Respects a cursor set from outside, otherwise falls back to the desired defaults

* Fixed embedded dialog host style (#2826) (#2829)

Using materialDesign:TransitionAssist.DisableTransitions="true" duration of animation for content cover opacity should be set to 0 to skip animations at all on close dialog host

* Fix for 2596 - Align left margin for error message on outlined TextBox with helper text left margin (#2820)

* Fix left margin for error message on outlined TextBox

The error message and helper text should have the same Margin.Left value for the outlined TextBox.

* Adjusted error text left margin for filled text fields

ComboBoxes, DatePicker and TimePicker all rely on the fields internally, and thus inherit these changes.

* Added UI test for the filled style as well

* Retain original bottom margin

I want to make as few changes as possible to the original style (margin)

* Let DataGridCell handle mouse events itself when already editing (#2821)

* Refactoring DataGridAssist input handling (#2824)

* Refactoring DataGridAssist mouse handling

Rather than handling the tunneling event and "pre-processing", it now handles the (already handled - by the cell) bubbling event to perform additional actions.

* Added template column example

Fixed issue where it would not switch to edit on click

Co-authored-by: Kevin Bost <kitokeboo@gmail.com>

* Initial PasswordBox "revealed" styles

There are still issues with the positioning of the floating hint when password is revealed but empty; the hint should float up, but it does not.

* Fixed alignment and initial wiring up

Floating hint still misplaced when password is empty and revealed

* Showcasing style in sample app + alignment of reveal button

Reveal button aligns with the "clear text" button for consistency. Although I think the "clear text" button is floating a little too high in 2 of the styles (but that is the same for the normal PasswordBox and TextBox styles)

* Showcase that password can still be set from XAML

* Change PasswordChanged event handler to weak event pattern

Avoid leaking memory by replacing the regular event handler with a weak event handler.

* Updated showcasing in the demo tool with a bit more stuff

* Fixed CA warning

* Added UI test for 3-way binding

* Resolving merge conflicts

* Resolving merge conflicts

* Fixing hint proxy issues with password box

* Moved samples and added SMTX wrapper

Also fixed a bug regarding the opacity of the PART_ClearButton which was not consistent across the styles.

* Apply cursor fix

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Material Design Service Account <mdixsa@outlook.com>
Co-authored-by: Andrey Nasonov <37536541+Erapchu@users.noreply.github.com>
Co-authored-by: Kevin Bost <kitokeboo@gmail.com>