Set vertical alignment of reveal button (#2873)
Setting to "Center" to align with PART_ClearButton. This threw me off, because the demo application overrides the style for PackIcon in Fields.xaml and explicitly sets VerticalAlignment=Center. The default however is Top, which is not correct. Instead of changing it directly on the PackIcon - which potentially is a bit too intrusive - I simply apply it directly on the PackIcon used in the reveal button.