Refactor PasswordBoxAssist.Password to use behaviors (issue 2930) (#2932)
* Rewrite current UI test and add new UI test to test failing scenario

* Refactor PasswordBoxAssist to use behavior

* Refactor PasswordBox reveal style TextBox focus- and text selection code

The old focus code did not work that well because keyboard focus was not really possible to set via the Style. With the introduction of behaviors, we can now improve that a lot.

* Undo wrong change in TestBase

* Whitespace commit to force re-run

* Increase delay in test

Test runs perfectly locally, so I don't understand why it fails in the pipeline

* Rollback of increased test delay

* Update MaterialDesignThemes.Wpf/Behaviors/PasswordBoxRevealTextBoxBehavior.cs

Co-authored-by: Kevin B <Keboo@users.noreply.github.com>

* Update MaterialDesignThemes.Wpf/Behaviors/PasswordBoxRevealTextBoxBehavior.cs

Co-authored-by: Kevin B <Keboo@users.noreply.github.com>

* Extract PasswordBoxBehavior into its own class

* Cache PropertyInfos and MethodInfos in PasswordBoxRevealTextBoxBehavior

* Selecting minimum required version of dependency and updating nuspec

I am unsure whether we can go down to an even older version - potentially supporting a broader audience.

* Convert Password binding to opt-in feature

Previously, the attached property was updated based on the PasswordChanged event regardless of whether the user had setup a binding or not. Now it is only wired up if there is a binding, or the "reveal" style has been applied (where it is a necessary evil).

* File scoped namespace

* Behavior classes renames and change from internal to public

* Use latest version of Xaml.Behaviors and update nuspec file

* Attempting fix of pipeline

* Adding screenshots for debugging

* Attempt at fixing failing UI test

* Adding more screenshots for debugging

* Activate test window on loaded event

* Yet another attempt at getting test window shown

* Removing debugging screenshots

Co-authored-by: Kevin B <Keboo@users.noreply.github.com>
Co-authored-by: Kevin Bost <kitokeboo@gmail.com>