Add special ComboBox popup handling for 90 degrees rotation
Popup placement was wrong in case a LayoutTransform was applied to rotate the ComboBox. The default WPF ComboBox does respect this, and thus so should MDIX.

To RotateTransfrom.Angle is compared against values of 90 and -90 (10 digit precision) to determine whether this is a rotation where the special handling should be applied.