DataGrid style cleanup (#3033)
* Introduce attached property for ComboBox column style - DataGrid

Similar to text and checkbox columns, the auto-generated style can now be set using an attached property.

* Introduce DataGridAssist.ApplyMaterialDesignColumnStyles attached property

The attached property, when set, inspects all columns of the data grid and applies MDIX styles if the default style has not been overridden. This works both for auto-generated and explicitly-defined columns.

* Fix nullability error when building in pipeline

* Fixed another nullable build error found in the pipeline

* Refactor DataGridAssist.ApplyMaterialDesignColumnStyles to use AP

The assist class now uses the attached properties associated with the corresponding DataGrid. Introduced a ColumnUpdater type and a private DP in order to have a reference to the DataGrid in the Columns_CollectionChanged() callback and to do proper housekeeping/cleanup to avoid memory leaks.

* Update MaterialDesignThemes.Wpf/DataGridAssist.cs

Co-authored-by: Kevin B <Keboo@users.noreply.github.com>

* Refactor to use local function to capture "dataGrid" variable

Removing the ColumnUpdater in place of a local function which captures the variable nicely. This removes the need for the ColumnUpdater type entirely, and removes the associated attached property.

Co-authored-by: Kevin B <Keboo@users.noreply.github.com>