Refactor SmartHint alignment (#3056)
* Refactor SmartHint alignment

Alignment now respects the HorizontalContentAlignment of the "owner"/"proxy". By default, the hint will float to the same alignment (left/center/right) as the HorizontalContentAlignment. Because I suspect there may be cases a user would want to float the hint differently that the resting position, I added an attached property HintAssist.FloatingHintHorizontalAlignment (inherited in the visual tree) which can be used to override the hint position when the hint is floated. By default, this property returns "Inherit" which means it follows the HorizontalContentAlignment, but it can also be set to Left/Center/Right/Stretch allowing explicit override of the default floating position.

* Update MaterialDesignThemes.Wpf/Converters/FloatingHintTextBlockMarginConverter.cs

Co-authored-by: Kevin B <Keboo@users.noreply.github.com>

---------

Co-authored-by: Kevin B <Keboo@users.noreply.github.com>