Workaround for Move() of items with expanded children
Since children get disconnected when moving the parent item this is a (hopefully temporary) workaround which simply collapses all children of the parent and its siblings before moving around in the collection. This is really not ideal behavior, and we should try to get it to work correctly.

I have left code in here which attempts to get it to work correctly, but it is not working yet.