Disable animation for "non transistion" states of ToggleButton
I may be wrong here, but I believe the <VisualState> storyboards should not have a duration. I believe these will be invoked when VisualStateManager.GoToState() is called with useTransitions=false. Thus I believe animations (durations) should only be used in the <VisualTransistion> storyboards which is invoked when VisualStateManager.GoToState() is called with useTransitions=true.