Ensure InteractionEllipse is not hit test visible
The interaction ellipse is used to provide visual feedback to the user when the radio button is being checked. However, it remains in the "large" state and since it was hit test visible, it captured mouse clicks outside of the bounds of the control, causing the radio button to be checked when it should not.