Post SmartHint adoption cleanup (#3540)
* Replicate TextBox style in AutoSuggestBox template

* Remove unused constants

* Drop default HintOffset to (0,0)

The HintOffset is now only used if explicitly set to a value other than the default; thus a default of (0,0) seems appropriate.

* Obsolete not needed assist classes

* Obsolete/delete unused converters

* Add delay to AutoSuggestBox test to make it less fragile.

* Ensure SUT is displayed in the middle of the window.
Filling the height of the host window is problematic when a Popup needs to be shown below the control. This causes the window to increase in size and does not really imitate real life usage.