Use tab animation even when tab has focusable content
Reverts the "destructive read" commit because the event may fire twice now, and removing the scroll direction would lead the second event to scroll to a wrong (i.e. not offset) position. So we live with the extra event (in cases where there is no focusable content; probably not the common use-case), and live with the AP/DP value surviving.